Move browser specific WebViewClient code into separate class (#281)
